pyproject.toml
[project]
name = "majodietagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=0.4.0",
    "fastapi>=0.115.0",
    "openai>=1.99.0",
    "pydantic>=2.11.0",
    "pypdf>=4.3.0",
    "python-multipart>=0.0.9",
    "psycopg>=3.2.0",
    "supabase>=2.27.2",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.0",
    "psycopg2-binary>=2.9.11",
]
